requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
openpyxl==3.1.2
APScheduler==3.10.4
